requirements.txt
flask>=3.0.0
gunicorn>=21.0.0
httpx>=0.25.0
python-dotenv>=1.0.0
onchaindb>=0.1.0
